Fix comparison warning

+++ libeasea/include/algorithms/moea/Cnsga-ii.h
@@ -143,7 +143,7 @@ template <typename TPtr, typename TIter> TIter Cnsga_ii<TIndividual, TRandom>::s
 {
         std::vector<TPtr> iFront(front.begin(), front.end());
         easea::shared::functions::setCrowdingDistance<TO>(iFront.begin(), iFront.end());
-        if (iFront.size() > std::distance(begin, end)) LOG_ERROR(errorCode::value, "Select Noncritical : Error of front size");
+        if (static_cast<long int>(iFront.size()) > std::distance(begin, end)) LOG_ERROR(errorCode::value, "Select Noncritical : Error of front size");
         TIter dest = begin;
         for (size_t i = 0; i < iFront.size(); ++i, ++dest)
                 *dest = *iFront[i];
@@ -157,7 +157,7 @@ template <typename TPtr, typename TIter> TIter Cnsga_ii<TIndividual, TRandom>::s
         easea::shared::functions::setCrowdingDistance<TO>(iFront.begin(), iFront.end());
         std::partial_sort(iFront.begin(), iFront.begin() + std::distance(begin, end), iFront.end()
                 , [](TPtr individual1, TPtr individual2)->bool{return individual1->m_crowdingDistance > individual2->m_crowdingDistance;});
-        if (iFront.size() < std::distance(begin, end)) LOG_ERROR(errorCode::value, "Select critical : Error of front size!");
+        if (static_cast<long int>(iFront.size()) < std::distance(begin, end)) LOG_ERROR(errorCode::value, "Select critical : Error of front size!");
         TIter dest = begin;
         for (size_t i = 0; dest != end; ++i, ++dest)
                 *dest = *iFront[i];
